requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
httpx==0.27.0
google-generativeai==0.3.2
tenacity==8.2.3
pydantic==2.6.4
aiofiles==23.2.1